fix bugs in hashing

+++ src/table.hpp
@@ -12,7 +12,7 @@ class Table {
    public:
     Table();
     void put(const Board& position, const TableEntry entry);
-    void put(const Board& position, Move move, int16_t eval, NodeType type, uint8_t depth);
+    void put(const Board& position, Move move, int32_t eval, NodeType type, uint8_t depth);
     std::optional<TableEntry> get(const Board& position);
     std::optional<TableEntry> get(const uint64_t hash);
     std::vector<Move> get_variation(Board& position);
