+++ src/table_entry.hpp
@@ -14,7 +14,7 @@ typedef enum NodeType {
 class TableEntry {
    public:
     TableEntry();
-    TableEntry(uint32_t hash, Move move, int16_t eval, NodeType type, uint8_t depth);
+    TableEntry(uint32_t hash, Move move, int32_t eval, NodeType type, uint8_t depth);
 
     // to string
     friend std::ostream& operator<<(std::ostream&, const TableEntry&);
@@ -22,7 +22,7 @@ class TableEntry {
     // getters
     uint32_t get_upper_hash() const;
     Move get_move() const;
-    int16_t get_eval() const;
+    int32_t get_eval() const;
     NodeType get_type() const;
     int8_t get_depth() const;
